fix: address review — drop contract doc, split query test, fix pyright
- Remove docs/FRAGMENT_QUERY_BIBLIOGRAPHY_CONTRACT.md; the bibliographyDocuments
  contract moves to the PR description (no new markdown in the repo).
- Split the limit-summary hydration tests into
  test_fragment_repository_query_summary.py, keeping both files under the
  250-line gate.
- Narrow the FragmentQueryResultSchema().load result with
  cast(FragmentQueryResult, ...) so task type-pyright is clean.
